+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+- URL provider got better at searching for annotation files
+
+### Fixed
+- genomes from ftp locations not working
+
 ## [0.9.2] - 2021-01-28
 
 ### Added